requirements.txt
# Python requirements
web3>=7.0.0
pynacl>=1.5.0
pytest>=9.0.0
eth-account>=0.11.0
python-dotenv>=1.0.0
